switched the doc of the button to doc.rs
